refactor(cache): extract filesystem info module
